fix(checker): preserve boolean literal display for enum targets
