+++ crates/tsz-checker/src/error_reporter/core/diagnostic_source.rs
@@ -2048,6 +2048,10 @@ impl<'a> CheckerState<'a> {
         {
             return Some(display);
         }
+        if let Some(display) = self.identifier_literal_initializer_source_display(expr_idx, target)
+        {
+            return Some(display);
+        }
         if let Some(display) = self.rebuilt_array_source_display(declared_type, target) {
             return Some(display);
         }
@@ -2179,87 +2183,6 @@ impl<'a> CheckerState<'a> {
         }
 
         replaced_object_member.then(|| displays.join(" & "))
-    }
-
-    pub(in crate::error_reporter) fn identifier_array_object_literal_source_display(
-        &mut self,
-        expr_idx: NodeIndex,
-        target: TypeId,
-    ) -> Option<String> {
-        let node = self.ctx.arena.get(expr_idx)?;
-        if node.kind != tsz_scanner::SyntaxKind::Identifier as u16 {
-            return None;
-        }
-        let sym_id = self.resolve_identifier_symbol(expr_idx)?;
-        let symbol = self.ctx.binder.get_symbol(sym_id)?;
-        if (symbol.flags & tsz_binder::symbol_flags::VARIABLE) == 0 {
-            return None;
-        }
-        let &decl_idx = symbol.declarations.first()?;
-        let decl = self.ctx.arena.get_variable_declaration_at(decl_idx)?;
-        let init_idx = decl.initializer.into_option()?;
-        let init_idx = self.ctx.arena.skip_parenthesized_and_assertions(init_idx);
-        let init_node = self.ctx.arena.get(init_idx)?;
-        if init_node.kind != syntax_kind_ext::ARRAY_LITERAL_EXPRESSION {
-            return None;
-        }
-        let literal = self.ctx.arena.get_literal_expr(init_node)?;
-        if literal.elements.nodes.is_empty() {
-            return None;
-        }
-
-        let mut ordered_names = Vec::new();
-        let mut property_values: Vec<Vec<TypeId>> = Vec::new();
-        for (element_index, &element_idx) in literal.elements.nodes.iter().enumerate() {
-            let element_node = self.ctx.arena.get(element_idx)?;
-            if element_node.kind != syntax_kind_ext::OBJECT_LITERAL_EXPRESSION {
-                return None;
-            }
-            let object = self.ctx.arena.get_literal_expr(element_node)?;
-            if element_index == 0 {
-                for &child_idx in &object.elements.nodes {
-                    let child = self.ctx.arena.get(child_idx)?;
-                    let prop = self.ctx.arena.get_property_assignment(child)?;
-                    let name = self.get_property_name(prop.name)?;
-                    ordered_names.push(name);
-                    property_values.push(vec![self.get_type_of_node(prop.initializer)]);
-                }
-                continue;
-            }
-
-            if object.elements.nodes.len() != ordered_names.len() {
-                return None;
-            }
-            for (prop_index, &child_idx) in object.elements.nodes.iter().enumerate() {
-                let child = self.ctx.arena.get(child_idx)?;
-                let prop = self.ctx.arena.get_property_assignment(child)?;
-                let name = self.get_property_name(prop.name)?;
-                if name != ordered_names[prop_index] {
-                    return None;
-                }
-                property_values[prop_index].push(self.get_type_of_node(prop.initializer));
-            }
-        }
-
-        let fields = ordered_names
-            .into_iter()
-            .zip(property_values)
-            .map(|(name, value_types)| {
-                let widened_types = value_types
-                    .into_iter()
-                    .map(|ty| self.widen_type_for_display(ty))
-                    .collect::<Vec<_>>();
-                let value_type = if widened_types.len() == 1 {
-                    widened_types[0]
-                } else {
-                    self.ctx.types.factory().union(widened_types)
-                };
-                let display = self.format_assignability_type_for_message(value_type, target);
-                format!("{name}: {display}")
-            })
-            .collect::<Vec<_>>()
-            .join("; ");
-        Some(format!("{{ {fields}; }}[]"))
     }
 
     pub(in crate::error_reporter) fn has_more_specific_diagnostic_at_span(
